refactor(datatypes): rename WriteProxyMessage func in ConcurrentWebSocket and move type Base64WSBuff
rename func WriteMessage in ConcurrentWebSocket  to WriteProxyMessage; move type
Base64WSBufferWriter to file ws_datatypes.go
