+++ wssocks/concurrent_websocket.go
@@ -25,7 +25,7 @@ func (wsc *ConcurrentWebSocket) WSClose() error {
 // write message to websocket, the data is fixed format @ProxyData
 // id: connection id
 // data: data to be written
-func (wsc *ConcurrentWebSocket) WriteMessage(id ksuid.KSUID, data []byte) error {
+func (wsc *ConcurrentWebSocket) WriteProxyMessage(id ksuid.KSUID, data []byte) error {
 	dataBase64 := base64.StdEncoding.EncodeToString(data)
 	jsonData := WebSocketMessage{
 		Id:   id.String(),
